Create Changed_Obj_Past31Days
We did not have source control so a lot of times we did not know if someone changed a program recently.  This would allow us to see if the object had been changed.
Each Union is over a different library so you will need to add/change/remove as needed.